move convert_bow_to_vw inside BOW class
